+++ src/sparse_hopnet/includes/SparseHN.hpp
@@ -9,7 +9,7 @@
 typedef struct s_iclamped
 {
     size_t      id;
-    double      val;
+    float       val;
 } t_iclamped;
 
 class SparseHN
@@ -31,5 +31,5 @@ class SparseHN
         /* Train using either series of clamped nodes or item-item similarity */
         void    train(std::vector<t_iclamped> &clamped);
         /* Infer next likeliest <out_size> items from a series of user interactions */
-        double  seq_energy(std::vector<t_iclamped> &clamped);
+        float  seq_energy(std::vector<t_iclamped> &clamped);
 };
