+++ src/sparse_hopnet/includes/Tensor.hpp
@@ -9,8 +9,9 @@
 
 # define FSIG "PK2\00"
 
-typedef std::map<size_t, std::vector<double>>   adj_t;
+typedef std::map<size_t, std::vector<float>>   adj_t;
 typedef std::deque<adj_t>                       tensor_t;
+
 
 class Tensor
 {
@@ -28,8 +29,8 @@ class Tensor
         size_t                  size();
         size_t                  getDepth();
 
-        std::vector<double>     &get(size_t x, size_t y);
-        std::vector<double>     &operator()(size_t x, size_t y);
+        std::vector<float>      &get(size_t x, size_t y);
+        std::vector<float>      &operator()(size_t x, size_t y);
         void                    save(std::string const &path);
         void                    load(std::string const &path);
 };
